backend/requirements.txt
# agentstack


colorama
tqdm
requests
python-dotenv
bs4

ollama
openai==0.28

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

fastapi
uvicorn
fastapi[all]
sqlalchemy
